requirements.txt
openai==0.28
Django==5.1.1
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
python-decouple==3.8
python-dotenv==1.0.1